move types to root of src to match common TS patterns
