+++ src/client/utils/sortSearchResults.ts
@@ -1,4 +1,4 @@
-import { InitialSearchResult, SearchResult } from "../../shared/types";
+import { InitialSearchResult, SearchResult } from "../../types";
 
 export function sortSearchResults(results: InitialSearchResult[]): void {
   results.forEach((item, index) => {
